@@ -1,22 +1,18 @@
-import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import {
   Component,
   ElementRef,
-  EventEmitter,
   HostListener,
-  Input,
-  Output,
-  SimpleChanges,
-  OnInit,
   ViewChild,
   inject,
 } from '@angular/core';
 import { Router } from '@angular/router';
-import { Pokemon } from '../../pokemon';
 import { Environment } from '../../environments';
 import { HeaderComponent } from '../header/header.component';
 import { LeftMenuComponent } from '../left-menu/left-menu.component';
-import { fromEvent, map, tap } from 'rxjs';
+import { Pokemon } from '../../pokemon';
+import { Data } from '../../data';
+
 /**
  * To Do:
  * Create a select element to show pokemons from different habitats
@@ -26,11 +22,12 @@ import { fromEvent, map, tap } from 'rxjs';
  */
 
 @Component({
-  selector: 'app-display',
-  standalone: true,
-  imports: [HttpClientModule, HeaderComponent, LeftMenuComponent],
-  templateUrl: './display.component.html',
-  styleUrl: './display.component.css',
+    selector: 'app-display',
+    imports: [
+        HeaderComponent, LeftMenuComponent
+    ],
+    templateUrl: './display.component.html',
+    styleUrl: './display.component.css'
 })
 export class DisplayComponent {
   constructor(private router: Router) {}
@@ -38,37 +35,31 @@ export class DisplayComponent {
   httpClient = inject(HttpClient);
   typeImg = Environment.typeImg;
   urlImg = Environment.urlImg;
-  data: any;
-  AmountPokemon = 20;
-  countPokemon!: number;
-  getCount() {
-    this.httpClient
-      .get(`${this.url}pokemon?limit=100000&offset=0`)
-      .subscribe((data: any) => {
-        this.countPokemon = data.count;
-      });
-  }
+  data?: any;
+  limit = 20;
+
+
   @ViewChild('scrollableDiv', { static: false }) scrollableDiv!: ElementRef;
   ngOnInit(): void {
-    this.getCount();
-    this.getData();
+    this.getData(this.limit);
   }
 
-  getData() {
+  getData(limit:number) {
     this.httpClient
-      .get(`${this.url}pokemon?limit=${this.AmountPokemon}&offset=0`)
+      .get(`${this.url}pokemon?limit=${limit}`)
       .subscribe((data: any) => {
         this.data = data.results;
       });
       
   }
 
-  onClick(id: number) {
-    this.router.navigate([`pokemon:${id}`]);
+  onClick(name:any ) {
+    
+    this.router.navigate([`pokemon/${name}`]);
   }
 
   getType() {
-    this.data.forEach((pokemon: any) => {
+    this.data?.results.forEach((pokemon: any) => {
       this.httpClient.get(pokemon.url).subscribe((data: any) => {
         pokemon.types = data.types;
       });
@@ -84,9 +75,9 @@ export class DisplayComponent {
     // Comprobar si el scroll está cerca del final (menos de 100px para llegar)
     const nearBottom = scrollTop + divHeight >= scrollHeight - 200;
 
-    if (nearBottom && this.AmountPokemon < this.countPokemon) {
-      this.AmountPokemon += 10; // Incrementar la variable AmountPokemon
-      this.getData(); // Llamada para obtener más datos
+    if (nearBottom ) {
+      this.limit += 10;
+      this.getData(this.limit); // Llamada para obtener más datos
     }
   }
 }
