@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-types',
-  standalone: true,
-  imports: [],
-  templateUrl: './types.component.html',
-  styleUrl: './types.component.css'
+    selector: 'app-types',
+    imports: [],
+    templateUrl: './types.component.html',
+    styleUrl: './types.component.css'
 })
 export class TypesComponent {
 
