@@ -1,6 +1,7 @@
 export const Environment={
     ApiUrl :'https://pokeapi.co/api/v2/',
     urlImg : `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/`,
-    typeImg:"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/types/generation-iv/diamond-pearl/"
+    typeImg: "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/types/generation-iv/diamond-pearl/",
+    urlSpecie: "https://pokeapi.co/api/v2/pokemon-species/"
 }
     
